+++ ktorfit-ksp/src/main/kotlin/de/jensklingenberg/ktorfit/model/annotations/HttpMethodAnnotation.kt
@@ -2,5 +2,5 @@ package de.jensklingenberg.ktorfit.model.annotations
 
 open class HttpMethodAnnotation(open val path: String, open val httpMethod: HttpMethod) : FunctionAnnotation()
 
-class CustomHttp(override val path: String, override val httpMethod: HttpMethod, val hasBody: Boolean = false) :
+class CustomHttp(override val path: String, override val httpMethod: HttpMethod, val hasBody: Boolean = false, val customValue : String) :
     HttpMethodAnnotation(path, httpMethod)