+++ ktorfit-ksp/src/main/kotlin/de/jensklingenberg/ktorfit/utils/Utils.kt
@@ -49,7 +49,7 @@ fun KSFunctionDeclaration.parseHTTPMethodAnno(name: String): HttpMethodAnnotatio
 
             if (name == "HTTP") {
                 this.getAnnotationsByType(de.jensklingenberg.ktorfit.http.HTTP::class).firstOrNull()?.let {
-                    CustomHttp(it.path, HttpMethod.valueOf(it.method), it.hasBody)
+                    CustomHttp(it.path,HttpMethod.CUSTOM , it.hasBody,it.method)
                 }
 
             } else {